Start adding validation

+++ api/src/collections/card/mutations.ts
@@ -7,6 +7,10 @@ interface DeleteCardArgs {
 const deleteCard: Resolver<DeleteCardArgs> = async (_, { id }, ctx) => {
 	const card = await ctx.prisma.card.findUnique({ where: { id } });
 
+	if (!card) {
+		throw new Error('Card not found');
+	}
+
 	if (card.userId === ctx.user.id) {
 		return ctx.prisma.card.delete({ where: { id } });
 	} else {
